clk: rockchip: add dynamically calculated PLL for rk3399
All RK3399 PLLs currently use a hardcoded table of frequences.
This is not sufficient for HDMI timing needs, where the monitor may
request all kinds of frequencies not found inside the table.

Implement an algorithm to dynamically calculate PLL parameters based
on a target frequency. So that rk3399_pll_rates could be removed in
the future.

A similar approach can be found in u-boot's clk_rk3399.c

Signed-off-by: Toby Chen <tobyc@nvidia.com>